Added a setter, and indent cleanup.
Added the setMultiplier( ) function which allows an override of the multiplier (from test.js by the user pressing the 'w' key) when the user chooses the Watopia map within Zwift.  For whatever reason Zwift decides to have a slope multiplier of 50 on Watopia, but all other maps have a multiplier of 100.